requirements.txt
anyio==4.4.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.6.2
exceptiongroup==1.2.1
flake8==7.0.0
flake8-docstrings==1.7.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mccabe==0.7.0
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
tornado==6.4
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
